+++ cli/src/todo_editor.rs
@@ -4,14 +4,13 @@
 
 use crate::parser::{format_datetime, parse_datetime};
 use aimcal_core::{Priority, Todo, TodoPatch, TodoStatus};
-use clap::ValueEnum;
-use crossterm::event::{self, Event, KeyCode};
+use crossterm::event::{self, Event, KeyCode, KeyEventKind};
 use ratatui::{
     prelude::*,
     symbols::border,
     widgets::{self, Block, Paragraph, block},
 };
-use std::error::Error;
+use std::{error::Error, rc::Rc};
 use unicode_width::UnicodeWidthStr;
 
 /// TUI editor for editing todos.
@@ -23,7 +22,15 @@ pub struct TodoEditor {
 
 impl TodoEditor {
     pub fn new() -> Self {
-        Self::new_with(Data::default())
+        Self::new_with(Data {
+            uid: "".to_string(),
+            description: "".to_string(),
+            due: "".to_string(),
+            percent_complete: "".to_string(),
+            priority: Priority::None,
+            status: TodoStatus::NeedsAction,
+            summary: "".to_string(),
+        })
     }
 
     pub fn from(todo: impl Todo) -> Self {
@@ -31,17 +38,16 @@ impl TodoEditor {
     }
 
     fn new_with(data: Data) -> Self {
-        Self {
-            store: Store::new(data),
-            fields: vec![
-                Component::Summary(FieldSummary::new(0)),
-                Component::Due(FieldDue::new(1)),
-                Component::PercentComplete(FieldPercentComplete::new(2)),
-                Component::Priority(FieldPriority::new(3)),
-                Component::Status(FieldStatus::new(4)),
-                Component::Description(FieldDescription::new(5)),
-            ],
-        }
+        let store = Store::new(data);
+        let fields = vec![
+            Component::Summary(FieldSummary::new(0)),
+            Component::Due(FieldDue::new(1)),
+            Component::PercentComplete(FieldPercentComplete::new(2)),
+            Component::Priority(FieldPriority::new(3, &store)),
+            Component::Status(FieldStatus::new(4)),
+            Component::Description(FieldDescription::new(5)),
+        ];
+        Self { store, fields }
     }
 
     pub fn run(mut self) -> Result<Option<TodoPatch>, Box<dyn Error>> {
@@ -78,7 +84,7 @@ impl TodoEditor {
 
     fn handle_event(&mut self, event: Event) -> Option<bool> {
         match event {
-            Event::Key(key) => self.handle_input(key.code),
+            Event::Key(key) if key.kind == KeyEventKind::Press => self.handle_input(key.code),
             _ => None,
         }
     }
@@ -108,7 +114,7 @@ impl TodoEditor {
             }
         };
 
-        if let Some(a) = self.fields.get(index) {
+        if let Some(a) = self.fields.get_mut(index) {
             a.update(&mut self.store, action)
         }
         None
@@ -122,7 +128,7 @@ impl TodoEditor {
         let areas = self.layout().split(area);
         let index = self.store.field_index;
         match (self.fields.get(index), areas.get(index)) {
-            (Some(field), Some(area)) => Some(field.get_cursor_position(&self.store, *area)),
+            (Some(field), Some(area)) => field.get_cursor_position(&self.store, *area),
             _ => None,
         }
     }
@@ -149,6 +155,7 @@ impl Widget for &TodoEditor {
             .title(title.centered())
             .title_bottom(instructions.centered())
             .border_set(border::ROUNDED)
+            .white()
             .render(area, buf);
 
         for (field, area) in self.fields.iter().zip(self.layout().split(area).iter()) {
@@ -198,12 +205,12 @@ impl Store {
         self.dirty.percent_complete = true;
     }
 
-    fn update_priority(&mut self, value: String) {
+    fn update_priority(&mut self, value: Priority) {
         self.data.priority = value;
         self.dirty.priority = true;
     }
 
-    fn update_status(&mut self, value: String) {
+    fn update_status(&mut self, value: TodoStatus) {
         self.data.status = value;
         self.dirty.status = true;
     }
@@ -222,7 +229,6 @@ impl Store {
     }
 
     fn submit(self) -> Result<TodoPatch, Box<dyn Error>> {
-        const IGNORE_CASE: bool = false;
         Ok(TodoPatch {
             uid: self.data.uid,
             description: match self.dirty.description {
@@ -239,14 +245,8 @@ impl Store {
                 true => Some(Some(self.data.percent_complete.parse::<u8>()?)),
                 false => None,
             },
-            priority: match self.dirty.priority {
-                true => Some(Priority::from_str(&self.data.priority, IGNORE_CASE)?),
-                false => None,
-            },
-            status: match self.dirty.status {
-                true => Some(TodoStatus::from_str(&self.data.status, IGNORE_CASE)?),
-                false => None,
-            },
+            priority: self.dirty.priority.then_some(self.data.priority),
+            status: self.dirty.status.then_some(self.data.status),
             summary: self.dirty.summary.then(|| self.data.summary.clone()),
         })
     }
@@ -261,14 +261,14 @@ enum Action {
     MoveRight,
 }
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 struct Data {
     uid: String,
     description: String,
     due: String,
     percent_complete: String,
-    priority: String,
-    status: String,
+    priority: Priority,
+    status: TodoStatus,
     summary: String,
 }
 
@@ -282,23 +282,8 @@ impl<T: Todo> From<T> for Data {
                 .percent_complete()
                 .map(|a| a.to_string())
                 .unwrap_or("".to_string()),
-            priority: match todo.priority() {
-                Priority::None => "none",
-                Priority::P1 => "1",
-                Priority::P2 => "2",
-                Priority::P3 => "3",
-                Priority::P4 => "4",
-                Priority::P5 => "3",
-                Priority::P6 => "6",
-                Priority::P7 => "7",
-                Priority::P8 => "8",
-                Priority::P9 => "9",
-            }
-            .to_string(),
-            status: todo
-                .status()
-                .map(|a| a.to_string())
-                .unwrap_or("".to_string()),
+            priority: todo.priority(),
+            status: todo.status().unwrap_or(TodoStatus::NeedsAction),
             summary: todo.summary().to_string(),
         }
     }
@@ -336,7 +321,7 @@ impl Component {
         }
     }
 
-    fn update(&self, store: &mut Store, action: Action) {
+    fn update(&mut self, store: &mut Store, action: Action) {
         match self {
             Component::Description(field) => field.update(store, action),
             Component::Due(field) => field.update(store, action),
@@ -347,7 +332,7 @@ impl Component {
         }
     }
 
-    fn get_cursor_position(&self, store: &Store, area: Rect) -> (u16, u16) {
+    fn get_cursor_position(&self, store: &Store, area: Rect) -> Option<(u16, u16)> {
         match self {
             Component::Description(field) => field.get_cursor_position(store, area),
             Component::Due(field) => field.get_cursor_position(store, area),
@@ -403,15 +388,21 @@ macro_rules! field_input {
                 self.input(store).render(area, buf);
             }
 
-            pub fn get_cursor_position(&self, store: &Store, area: Rect) -> (u16, u16) {
-                self.input(store)
-                    .get_cursor_position(area, store.character_index)
+            pub fn get_cursor_position(&self, store: &Store, area: Rect) -> Option<(u16, u16)> {
+                Some(
+                    self.input(store)
+                        .get_cursor_position(area, store.character_index),
+                )
             }
 
             fn input<'a>(&self, store: &'a Store) -> Input<'a> {
                 let value = store.data.$field.as_str();
                 let active = store.field_index == self.index;
-                Input::new($title, value, active)
+                Input {
+                    title: $title,
+                    value,
+                    active,
+                }
             }
         }
     };
@@ -430,26 +421,179 @@ field_input!(
     percent_complete,
     update_percent_complete
 );
-field_input!(FieldPriority, "Priority", priority, update_priority);
-field_input!(FieldStatus, "Status", status, update_status);
 field_input!(FieldSummary, "Summary", summary, update_summary);
 
-struct Input<'a> {
-    title: &'a str,
-    value: &'a str,
-    active: bool,
-}
+trait FieldSwitch {
+    type T: Eq;
 
-impl<'a> Input<'a> {
-    fn new(title: &'a str, value: &'a str, active: bool) -> Self {
-        Self {
-            title,
-            value,
-            active,
+    fn get_title() -> &'static str;
+    fn get_value(store: &Store) -> Self::T;
+    fn update_value(store: &mut Store, value: &Self::T);
+
+    fn index(&self) -> usize;
+    fn values(&self) -> &Vec<Self::T>;
+    fn options(&self) -> &Vec<String>;
+
+    fn update(&mut self, store: &mut Store, action: Action) {
+        match action {
+            Action::MoveTo => {
+                store.update_field_index(self.index());
+                store.update_character_index(self.selected(store));
+            }
+            Action::MoveLeft => {
+                let prev = (self.selected(store) + self.values().len() - 1) % self.values().len();
+                store.update_character_index(prev);
+                if let Some(a) = self.values().get(prev) {
+                    Self::update_value(store, a);
+                }
+            }
+            Action::MoveRight => {
+                let next = (self.selected(store) + 1) % self.values().len();
+                store.update_character_index(next);
+                if let Some(a) = self.values().get(next) {
+                    Self::update_value(store, a);
+                }
+            }
+            _ => {}
         }
     }
 
-    fn get_cursor_position(&self, area: Rect, character_index: usize) -> (u16, u16) {
+    fn render(&self, store: &Store, area: Rect, buf: &mut Buffer) {
+        self.switch(store).render(area, buf);
+    }
+
+    fn get_cursor_position(&self, store: &Store, area: Rect) -> Option<(u16, u16)> {
+        self.switch(store)
+            .get_cursor_position(area, store.character_index)
+    }
+
+    fn switch<'a>(&'a self, store: &'a Store) -> Switch<'a> {
+        Switch {
+            title: Self::get_title(),
+            values: self.options(),
+            active: store.field_index == self.index(),
+            selected: self.selected(store),
+        }
+    }
+
+    fn selected(&self, store: &Store) -> usize {
+        let v = Self::get_value(store);
+        self.values().iter().position(|s| s == &v).unwrap_or(0)
+    }
+}
+
+#[derive(Debug)]
+struct FieldPriority {
+    index: usize,
+    values: Vec<Priority>,
+    options: Vec<String>,
+}
+
+impl FieldPriority {
+    pub fn new(index: usize, store: &Store) -> Self {
+        let verbose = Self::need_verbose(&store.data.priority);
+        let (values, options) = Self::get_value_options(verbose);
+        Self {
+            index,
+            values,
+            options,
+        }
+    }
+
+    fn need_verbose(priority: &Priority) -> bool {
+        use Priority::*;
+        matches!(priority, P1 | P3 | P4 | P6 | P7 | P9)
+    }
+
+    fn get_value_options(verbose: bool) -> (Vec<Priority>, Vec<String>) {
+        use Priority::*;
+        let values = match verbose {
+            true => vec![P1, P2, P3, P4, P5, P6, P7, P8, P9, None],
+            false => vec![P2, P5, P8, None],
+        };
+
+        let options = values
+            .iter()
+            .map(|a| Self::fmt(a, verbose).to_string())
+            .collect();
+
+        (values, options)
+    }
+
+    fn fmt(priority: &Priority, verbose: bool) -> &'static str {
+        match priority {
+            Priority::P2 if !verbose => "high",
+            Priority::P5 if !verbose => "mid",
+            Priority::P8 if !verbose => "low",
+            Priority::None => "none",
+            Priority::P1 => "1",
+            Priority::P2 => "2",
+            Priority::P3 => "3",
+            Priority::P4 => "4",
+            Priority::P5 => "5",
+            Priority::P6 => "6",
+            Priority::P7 => "7",
+            Priority::P8 => "8",
+            Priority::P9 => "9",
+        }
+    }
+}
+
+#[rustfmt::skip]
+impl FieldSwitch for FieldPriority {
+    type T = Priority;
+
+    fn get_title() -> &'static str                      { "Priority" }
+    fn get_value(store: &Store) -> Self::T              { store.data.priority }
+    fn update_value(store: &mut Store, value: &Self::T) { store.update_priority(*value); }
+
+    fn index(&self) -> usize          { self.index }
+    fn values(&self) -> &Vec<Self::T> { &self.values }
+    fn options(&self) -> &Vec<String> { &self.options }
+}
+
+#[derive(Debug)]
+struct FieldStatus {
+    index: usize,
+    values: Vec<TodoStatus>,
+    options: Vec<String>,
+}
+
+impl FieldStatus {
+    pub fn new(index: usize) -> Self {
+        use TodoStatus::*;
+        let values = vec![NeedsAction, Completed, InProcess, Cancelled];
+        let options = values.iter().map(ToString::to_string).collect();
+        Self {
+            index,
+            values,
+            options,
+        }
+    }
+}
+
+#[rustfmt::skip]
+impl FieldSwitch for FieldStatus {
+    type T = TodoStatus;
+
+    fn get_title() -> &'static str                      { "Status" }
+    fn get_value(store: &Store) -> Self::T              { store.data.status }
+    fn update_value(store: &mut Store, value: &Self::T) { store.update_status(*value); }
+
+    fn index(&self) -> usize          { self.index }
+    fn values(&self) -> &Vec<Self::T> { &self.values }
+    fn options(&self) -> &Vec<String> { &self.options }
+
+}
+
+struct Input<'a> {
+    pub title: &'a str,
+    pub value: &'a str,
+    pub active: bool,
+}
+
+impl<'a> Input<'a> {
+    pub fn get_cursor_position(&self, area: Rect, character_index: usize) -> (u16, u16) {
         let index = character_index.min(self.value.len());
         let width = self.value[0..index].width();
         let x = area.x + (width as u16) + 2; // border 1 + padding 1
@@ -460,19 +604,68 @@ impl<'a> Input<'a> {
 
 impl Widget for &Input<'_> {
     fn render(self, area: Rect, buf: &mut Buffer) {
-        let block = Block::bordered()
-            .border_set(border::ROUNDED)
-            .borders(widgets::Borders::ALL)
-            .padding(widgets::Padding::horizontal(1))
-            .title_position(block::Position::Top)
-            .title(format!(" {} ", self.title).bold());
-
-        let block = match self.active {
-            true => block.blue(),
-            false => block.gray(),
-        };
-
+        let block = title_block(self.title, self.active);
         Paragraph::new(self.value).block(block).render(area, buf);
     }
 }
 
+struct Switch<'a> {
+    pub title: &'a str,
+    pub values: &'a Vec<String>,
+    pub active: bool,
+    pub selected: usize,
+}
+
+impl<'a> Switch<'a> {
+    pub fn get_cursor_position(&self, area: Rect, index: usize) -> Option<(u16, u16)> {
+        let (_, inners) = self.split(area);
+        inners.get(index).map(|area| {
+            let x = area.x + 2; // 2 = padding (1) + active marker [ (1)
+            (x, area.y)
+        })
+    }
+
+    fn split(&self, area: Rect) -> (Block, Rc<[Rect]>) {
+        let outer_block = title_block(self.title, self.active).white();
+        let inner = outer_block.inner(area);
+        let inner_blocks = self.layout().split(inner);
+        (outer_block, inner_blocks)
+    }
+
+    fn layout(&self) -> Layout {
+        let constraints = self
+            .values
+            .iter()
+            // 6 = border left (1) + active marker [ ] (3) + space (1) + border right (1)
+            .map(|s| Constraint::Min((6 + s.width()) as u16));
+
+        Layout::horizontal(constraints)
+    }
+}
+
+impl Widget for &Switch<'_> {
+    fn render(self, area: Rect, buf: &mut Buffer) {
+        let (outer, inners) = self.split(area);
+        outer.render(area, buf);
+
+        for (i, (value, area)) in self.values.iter().zip(inners.iter()).enumerate() {
+            let label = format!("[{}] {}", if self.selected == i { 'x' } else { ' ' }, value);
+            let inner_block = Block::default().padding(block::Padding::horizontal(1));
+            Paragraph::new(label).block(inner_block).render(*area, buf);
+        }
+    }
+}
+
+fn title_block(title: &str, active: bool) -> Block {
+    let block = Block::bordered()
+        .border_set(border::ROUNDED)
+        .borders(widgets::Borders::ALL)
+        .padding(widgets::Padding::horizontal(1))
+        .title_position(block::Position::Top)
+        .title(format!(" {title} ").bold());
+
+    match active {
+        true => block.blue(),
+        false => block.white(),
+    }
+}