+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add TUI switch for status and priority
+
 ## [v0.2.1] - 2025-07-27
 
 ### Added